Documented the Struct class and changed some parameters.
Instead of Collections, the static factory methods now take Iterables
to indicate, that the given argument will not be modified my the class.
The Getter of the "attributes" will now also return a Stream instead
of a Collection. This allows easier use and also indicates, that the
attributes may not be modified after construction.